Update to 1.21.1
There are some *changes* changes, such as the removal of the "handleBlockBreakAck" packet handler as the packet no longer contains the block position. It probably doesn't matter though as block updates should be handled by the other packets.

Another thing is the TODO now in the Cartography class as I am not certain that the behaviour is the same. May need some testing.

Also removed the "xaeros-minimap" dependency as it's not used.

There's a problem with the Forge client that I haven't been able to fix. It errors saying that the same package is being exported: "Modules MapSync.dev and generated_a20f2c1 export package gjum.minecraft.mapsync.common.data to module mixin.synthetic". No idea how to fix. Tempted to just switch to NeoForge.